Rendering rework

+++ core/src/com/mygdx/game/gui/InventorySlot.java
@@ -12,7 +12,7 @@ import com.mygdx.game.item.Item;
 
 public class InventorySlot extends Widget {
 
-    public static final float SIZE = 36;
+    public static final int SIZE = 36;
     public static final float OFFSET = 2;
 
     private Skin skin;
